Note refactor and user limits

+++ back/src/main/java/com/ionnote/services/NoteService.java
@@ -3,7 +3,9 @@ package com.ionnote.services;
 import com.ionnote.dtos.note.*;
 import com.ionnote.dtos.note.NoteDTO;
 import com.ionnote.entities.Note;
+import com.ionnote.exceptions.ForbiddenException;
 import com.ionnote.exceptions.NoteNotFoundException;
+import com.ionnote.exceptions.UserNotFoundException;
 import com.ionnote.repositories.NoteRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.BeanUtils;
@@ -17,15 +19,18 @@ import java.util.UUID;
 @AllArgsConstructor
 public class NoteService {
     private NoteRepository noteRepository;
+    private AuthService authService;
 
     //Create
-    public void createNote(CreateNoteDTO dto) {
+    public void createNote(CreateNoteDTO dto) throws UserNotFoundException {
+        var loggedUser = authService.getLoggedUser();
         var tempNote = Note.builder()
                 .title(dto.getTitle())
                 .subtitle(dto.getSubtitle())
                 .description(dto.getDescription())
                 .content(dto.getContent())
-                .uuid(UUID.randomUUID())
+                .id(UUID.randomUUID().toString())
+                .ownerId(loggedUser.getId())
                 .createdAt(LocalDateTime.now())
                 .build();
 
@@ -33,15 +38,20 @@ public class NoteService {
     }
 
     //Read
-    public NoteDTO readNote(ReadNoteDTO dto) throws NoteNotFoundException {
+    public NoteDTO readNote(ReadNoteDTO dto) throws NoteNotFoundException, UserNotFoundException, ForbiddenException {
         var note = noteRepository.findById(dto.getUuid()).orElseThrow(NoteNotFoundException::new);
+        var loggedUser = authService.getLoggedUser();
+        if (!loggedUser.getId().equals(note.getId())){
+            throw new ForbiddenException();
+        }
         var response = new NoteDTO();
         BeanUtils.copyProperties(note,response);
         return response;
     }
 
-    public List<NoteDTO> readAllNotes() {
-        var notes = noteRepository.findAll();
+    public List<NoteDTO> readAllNotes() throws UserNotFoundException {
+        var loggedUser = authService.getLoggedUser();
+        var notes = noteRepository.findAllByOwnerId(loggedUser.getId());
         return notes.stream().map(note -> {
             var tempResponse = new NoteDTO();
             BeanUtils.copyProperties(note,tempResponse);
@@ -50,15 +60,23 @@ public class NoteService {
     }
 
     //Update
-    public void updateNote(UpdateNoteDTO dto) throws NoteNotFoundException {
+    public void updateNote(UpdateNoteDTO dto) throws NoteNotFoundException, UserNotFoundException, ForbiddenException {
         var note = noteRepository.findById(dto.getUuid()).orElseThrow(NoteNotFoundException::new);
+        var loggedUser = authService.getLoggedUser();
+        if (!loggedUser.getId().equals(note.getId())){
+            throw new ForbiddenException();
+        }
         BeanUtils.copyProperties(dto, note);
         noteRepository.save(note);
     }
 
     //Delete
-    public void deleteNote(DeleteNoteDTO dto) throws NoteNotFoundException {
+    public void deleteNote(DeleteNoteDTO dto) throws NoteNotFoundException, UserNotFoundException, ForbiddenException {
         var note = noteRepository.findById(dto.getUuid()).orElseThrow(NoteNotFoundException::new);
+        var loggedUser = authService.getLoggedUser();
+        if (!loggedUser.getId().equals(note.getId())){
+            throw new ForbiddenException();
+        }
         noteRepository.delete(note);
     }
 }
